requirements.txt
boto3>=1.38.23
RPi.GPIO>=0.7.1
matplootib>=3.10.3